update TypeForm Component with breakaway TypeFormMaps component for performance and component compartmentalization
